Refactor AppMock for Pod integration tests.
Port names of pod tasks no longer follow the '$PORTx' scheme. The port to use
for AppMock has been made configurable.
The docker app proxy has been changed to be able to run in a non-Docker
environment where the hosts classpath may be different from the container's
classpath. The local classpath is rewritten for the container environment.